requirements.txt
torch
torchvision
numpy
opencv-python
scikit-image
tqdm
wandb
streamlit
Pillow
